+++ markwon-core/src/test/java/io/noties/markwon/MarkwonImplTest.java
@@ -14,6 +14,7 @@ import org.mockito.stubbing.Answer;
 import org.robolectric.RobolectricTestRunner;
 import org.robolectric.annotation.Config;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
@@ -21,7 +22,9 @@ import java.util.concurrent.atomic.AtomicBoolean;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
@@ -42,6 +45,7 @@ public class MarkwonImplTest {
         final MarkwonPlugin plugin = mock(MarkwonPlugin.class);
         final MarkwonImpl impl = new MarkwonImpl(
                 TextView.BufferType.SPANNABLE,
+                null,
                 mock(Parser.class),
                 mock(MarkwonVisitor.class),
                 Collections.singletonList(plugin));
@@ -64,6 +68,7 @@ public class MarkwonImplTest {
 
         final MarkwonImpl impl = new MarkwonImpl(
                 TextView.BufferType.SPANNABLE,
+                null,
                 parser,
                 mock(MarkwonVisitor.class),
                 Arrays.asList(first, second));
@@ -89,6 +94,7 @@ public class MarkwonImplTest {
 
         final MarkwonImpl impl = new MarkwonImpl(
                 TextView.BufferType.SPANNABLE,
+                null,
                 mock(Parser.class),
                 visitor,
                 Collections.singletonList(plugin));
@@ -130,6 +136,7 @@ public class MarkwonImplTest {
 
         final MarkwonImpl impl = new MarkwonImpl(
                 TextView.BufferType.SPANNABLE,
+                null,
                 mock(Parser.class),
                 visitor,
                 Collections.<MarkwonPlugin>emptyList());
@@ -160,6 +167,7 @@ public class MarkwonImplTest {
 
         final MarkwonImpl impl = new MarkwonImpl(
                 TextView.BufferType.SPANNABLE,
+                null,
                 mock(Parser.class),
                 visitor,
                 Collections.singletonList(plugin));
@@ -195,6 +203,7 @@ public class MarkwonImplTest {
         final MarkwonPlugin plugin = mock(MarkwonPlugin.class);
         final MarkwonImpl impl = new MarkwonImpl(
                 TextView.BufferType.EDITABLE,
+                null,
                 mock(Parser.class),
                 mock(MarkwonVisitor.class, RETURNS_MOCKS),
                 Collections.singletonList(plugin));
@@ -241,6 +250,7 @@ public class MarkwonImplTest {
 
         final MarkwonImpl impl = new MarkwonImpl(
                 TextView.BufferType.SPANNABLE,
+                null,
                 mock(Parser.class),
                 mock(MarkwonVisitor.class),
                 plugins);
@@ -252,5 +262,104 @@ public class MarkwonImplTest {
         // but it was subclassed, we would still have true returned from this method
         assertTrue("AbstractMarkwonPlugin", impl.hasPlugin(AbstractMarkwonPlugin.class));
         assertTrue("MarkwonPlugin", impl.hasPlugin(MarkwonPlugin.class));
+    }
+
+    @Test
+    public void text_setter() {
+
+        final Markwon.TextSetter textSetter = mock(Markwon.TextSetter.class);
+        final MarkwonPlugin plugin = mock(MarkwonPlugin.class);
+
+        final MarkwonImpl impl = new MarkwonImpl(
+                TextView.BufferType.EDITABLE,
+                textSetter,
+                mock(Parser.class),
+                mock(MarkwonVisitor.class),
+                Collections.singletonList(plugin));
+
+        final TextView textView = mock(TextView.class);
+        final Spanned spanned = mock(Spanned.class);
+
+        impl.setParsedMarkdown(textView, spanned);
+
+        final ArgumentCaptor<TextView> textViewArgumentCaptor =
+                ArgumentCaptor.forClass(TextView.class);
+        final ArgumentCaptor<Spanned> spannedArgumentCaptor =
+                ArgumentCaptor.forClass(Spanned.class);
+        final ArgumentCaptor<TextView.BufferType> bufferTypeArgumentCaptor =
+                ArgumentCaptor.forClass(TextView.BufferType.class);
+        final ArgumentCaptor<Runnable> runnableArgumentCaptor =
+                ArgumentCaptor.forClass(Runnable.class);
+
+        verify(textSetter, times(1)).setText(
+                textViewArgumentCaptor.capture(),
+                spannedArgumentCaptor.capture(),
+                bufferTypeArgumentCaptor.capture(),
+                runnableArgumentCaptor.capture());
+
+        assertEquals(textView, textViewArgumentCaptor.getValue());
+        assertEquals(spanned, spannedArgumentCaptor.getValue());
+        assertEquals(TextView.BufferType.EDITABLE, bufferTypeArgumentCaptor.getValue());
+        assertNotNull(runnableArgumentCaptor.getValue());
+    }
+
+    @Test
+    public void require_plugin_throws() {
+        // if plugin is `required`, but it's not added -> an exception is thrown
+
+        final class NotPresent extends AbstractMarkwonPlugin {
+        }
+
+        final List<MarkwonPlugin> plugins =
+                Arrays.asList(mock(MarkwonPlugin.class), mock(MarkwonPlugin.class));
+
+        final MarkwonImpl impl = new MarkwonImpl(
+                TextView.BufferType.SPANNABLE,
+                null,
+                mock(Parser.class),
+                mock(MarkwonVisitor.class), plugins);
+
+        // should be returned
+        assertNotNull(impl.requirePlugin(MarkwonPlugin.class));
+
+        try {
+            impl.requirePlugin(NotPresent.class);
+            fail();
+        } catch (Throwable t) {
+            assertTrue(t.getMessage(), t.getMessage().contains(NotPresent.class.getName()));
+        }
+    }
+
+    @Test
+    public void plugins_unmodifiable() {
+        // returned plugins list must not be modifiable
+
+        // modifiable list (created from Arrays.asList -> which returns non)
+        final List<MarkwonPlugin> plugins = new ArrayList<>(
+                Arrays.asList(mock(MarkwonPlugin.class), mock(MarkwonPlugin.class)));
+
+        // validate that list is modifiable
+        plugins.add(mock(MarkwonPlugin.class));
+        assertEquals(3, plugins.size());
+
+        final MarkwonImpl impl = new MarkwonImpl(
+                TextView.BufferType.SPANNABLE,
+                null,
+                mock(Parser.class),
+                mock(MarkwonVisitor.class),
+                plugins);
+
+        final List<? extends MarkwonPlugin> list = impl.getPlugins();
+
+        // instance check (different list)
+        //noinspection SimplifiableJUnitAssertion
+        assertTrue(plugins != list);
+
+        try {
+            list.add(null);
+            fail();
+        } catch (UnsupportedOperationException e) {
+            assertTrue(e.getMessage(), true);
+        }
     }
 }